+++ CHANGELOG.md
@@ -5,10 +5,16 @@
 ### 1.1 ###
 
 * Added `CHANGELOG.md`
+* Updated `README.md`
+* Added `LICENSE.md` (GNU General Public License, version 2)
 * When using `--check_accuracy`, PCAViz also reports the:
   1) number of components retained to achieve the target cumulative variance
   2) size of output JSON file
   3) approximate size of output JSON file if it were compressed with gzip
+* Created separate ZIP files for the JavaScript Interpreter, Python
+  Compressor, and WordPress plugin so users don't have to download the whole
+  repository
+* Improved build process (primarily for Durrant-Lab use)
 
 ### 1.0 ###